Perform a long-overdue update to the CoC
We have always referred to this as the Kinetic Cafe code of conduct, but
that company no longer exists. We use github.com/KineticCafe to host our
open source repositories, but this is the Kinetic Commerce code of
conduct which will be enforced by employees of Kinetic Commerce.

While updating this, we changed some of the wording to try to better
reflect that any enforcement will be at a *repository* ("project") level
first and foremost, but may expand to participation in all KineticCafe
repositories. We are not an OSS community like Mozilla, Debian, or even
the former Twitter open source community where there were employees
whose full-time job was to manage the open source contributions of their
employees and external contributors (often requiring the acceptance of
contributor licensing agreements).

We have had some contributions in the past to our open source projects,
but very few.

Review of the most recent version of the Contributor Covenant offers
some useful wording updates to be *clearer* on acceptable and
unacceptable behaviours as well as enforcement stages. These changes are
drastic enough that it was worthwhile marking a major version bump.